+++ chula/json.py
@@ -2,15 +2,47 @@
 Wrapper to make it easy to switch from one json library to another
 """
 
+# Python imports
+import sys
+
+# Chula imports
 from chula import error
 
-try:
-    import simplejson
-except:
-    try:
-        from django.utils import simplejson
-    except:
-        raise error.MissingDependencyError('Simplejson')
+def _official():
+    sys.path.pop(0) # The python provided lib has a namespace clash
+    import json
+    return json
 
-decode = simplejson.loads
-encode = simplejson.dumps
+def _simplejson():
+    import simplejson
+    return simplejson
+
+def _django():
+    from django.utils import simplejson
+    return simplejson
+
+# Fetch the [most] preferred json provider
+json_provider = None
+for provider in [_official, _simplejson, _django]:
+    try:
+        json_provider = provider()
+        break
+    except ImportError:
+        pass
+
+# Make sure we found a provider
+if json_provider is None:
+    msg = 'Simplejson, or Python >=2.6'
+    raise error.MissingDependencyError(msg)
+
+# Expose the two methods we care about
+decode = json_provider.loads
+encode = json_provider.dumps
+
+if __name__ == '__main__':
+    data = {'abc':'123', 'foobar':45}
+    encoded = encode(data)
+    decoded = decode(encoded)
+    print('Provider:', json_provider)
+    print('Encoded:', encoded)
+    print('Decoded:', decoded)